pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-chatbot-portfolio"
version = "0.1.0"
description = "Production-style RAG chatbot portfolio project"
requires-python = ">=3.10"
dependencies = [
  "fastapi==0.115.2",
  "uvicorn[standard]==0.30.6",
  "streamlit==1.39.0",
  "pydantic==2.9.2",
  "pydantic-settings==2.5.2",
  "python-multipart==0.0.12",
  "httpx==0.27.2",
  "chromadb==0.5.11",
  "sentence-transformers==3.2.1",
  "rank-bm25==0.2.2",
  "PyMuPDF==1.24.10",
  "python-docx==1.1.2",
  "tiktoken==0.8.0",
  "numpy==1.26.4",
  "pandas==2.2.3",
  "scikit-learn==1.5.2",
  "orjson==3.10.7",
  "python-dotenv==1.0.1",
  "ragas==0.2.3",
  "datasets==3.0.1",
  "pytest==8.3.3",
  "pytest-asyncio==0.24.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"
